services/_shared/audit_common/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "audit_common"
version = "0.1.0"
description = "Shared db / auth / logging / errors for VIA backend services"
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "VIA Audit team" }]
dependencies = [
    "asyncpg>=0.29",
    "fastapi>=0.110",
    "PyJWT>=2.8",
]

[tool.setuptools]
packages = ["audit_common"]
package-dir = { audit_common = "." }

[tool.mypy]
strict = true
ignore_missing_imports = false
